ログイン処理時に、firebaseにアカウントを登録する処理を追加

+++ lib/pages/profile.dart
@@ -176,7 +176,9 @@ class _ProfileState extends State<Profile> {
         .document(currentUserId)
         .get()
         .then((doc) {
-      doc.reference.delete();
+      if (doc.exists) {
+        doc.reference.delete();
+      }
     });
     // remove following
     followingRef
